+++ src/assets/chakraTheme.ts
@@ -1,0 +1,15 @@
+import { extendTheme } from '@chakra-ui/react';
+
+export default extendTheme({
+    styles: {
+        global: {
+            body: {
+                bg: 'gray.100',
+            },
+        },
+    },
+    fonts: {
+        heading: 'Inter',
+        body: 'Inter',
+    },
+});