+++ src/assets/chakraTheme/index.ts
@@ -1,9 +1,0 @@
-import { extendTheme } from '@chakra-ui/react';
-
-import styles from '@/assets/chakraTheme/styles';
-
-const overrides = {
-    styles,
-};
-
-export default extendTheme(overrides);